requirements.txt
torch
torchvision
matplotlib
scikit-learn
tqdm
pillow
numpy
pandas
streamlit
boto3
